radarsim/pyproject.toml
[project]
name = "radarsim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
]
